# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the StudentLink Server API ",
        "version": "0.0.1",
        "title": "StudentLink API"
    },
    "tags": [
        {
            "name": "Courses",
            "description": "Endpoints related to viewing and manipulating Course data"
        },
        {
            "name": "Users",
            "description": "Endpoints related to viewing and manipulating User data"
        }
    ],
    "schemes": ["http"],
    "paths": {
        "/course/{courseId}": {
            "get": {
                "tags": ["Courses"],
                "summary": "Get a Course object by ID",
                "description": "",
                "operationId": "",
                "produces": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "courseId",
                        "description": "The ID of the Course to return",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/course": {
            "post": {
                "tags": ["Courses"],
                "summary": "Create a Course by providing a course code and course section",
                "description": "",
                "operationId": "",
                "produces": [],
                "parameters": [
                    {
                        "in": "body",
                        "name": "courseCode",
                        "description": "The course code, e.g. 'CPEN321'",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "courseSection",
                        "description": "The course section, e.g. 'L1A'",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/user/{userId}": {
            "get": {
                "tags": ["Users"],
                "summary": "Get a User object by ID",
                "description": "",
                "operationId": "",
                "produces": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "description": "The ID of the User to return",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/user/{userId}/": {
            "patch": {
                "tags": ["Users"],
                "summary": "Update a User object by ID",
                "description": "",
                "operationId": "",
                "produces": []
            }
        },
        "/user": {
            "post": {
                "tags": ["Users"],
                "summary": "Create a User by providing a first name and last name",
                "description": "",
                "operationId": "",
                "produces": [],
                "parameters": [
                    {
                        "in": "body",
                        "name": "firstName",
                        "description": "The user's first name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "lastName",
                        "description": "The user's last name",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/group/{groupId}": {
            "get": {
                "tags": ["Group"],
                "summary": "Get a Group object by ID",
                "description": "",
                "operationId": "",
                "produces": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "groupId",
                        "description": "The ID of the Group to return",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/group/{groupId}/": {
            "patch": {
                "tags": ["Group"],
                "summary": "Update a Group object by ID",
                "description": "",
                "operationId": "",
                "produces": []
            }
        },
        "/group": {
            "post": {
                "tags": ["Group"],
                "summary": "Create a Group by providing a User's userId, schedule, and registered courses' courseIds",
                "description": "",
                "operationId": "",
                "produces": [],
                "parameters": [
                    {
                        "in": "body",
                        "name": "members",
                        "description": "The founding User's userId",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "body",
                        "name": "courses",
                        "description": "The founding User's courses",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "body",
                        "name": "meeting_times",
                        "description": "The founding User's schedule",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                ]
            }
        }
    }
}